+++ include/GRMustacheVersion.h
@@ -26,19 +26,19 @@
  * 
  * @since v1.0
  */
-#define GRMUSTACHE_MAJOR_VERSION 4
+#define GRMUSTACHE_MAJOR_VERSION 5
 
 /**
  * The minor component of GRMustache version
  * 
  * @since v1.0
  */
-#define GRMUSTACHE_MINOR_VERSION 3
+#define GRMUSTACHE_MINOR_VERSION 0
 
 /**
  * The patch-level component of GRMustache version
  * 
  * @since v1.0
  */
-#define GRMUSTACHE_PATCH_VERSION 4
+#define GRMUSTACHE_PATCH_VERSION 0